PLANET-7849: Refactor header.js file
- Move elements to another file
